carrito ya casi

+++ TiendaMascotas/TiendaMascotas/Controllers/ProductosController.cs
@@ -79,7 +79,7 @@ namespace TiendaMascotas.Controllers
         //[FiltroSesion]
         //Parte 1 - Con AJAX, se llama desde ~/assets/jsProductos/productos.js que eventualmente se llama desde la vista ProductosVenta.cshtml en el Eliminar
         //Parte 2 - Una vez se recibe respuesta del API/ProductosModel/EliminarProducto se dice que se eliminó correctamente (devuelve 'Ok')
-        public ActionResult EliminarProducto(long id)
+        public ActionResult EliminarProducto(int id)
         {
             model.EliminarProducto(id);
             return Json("Ok", JsonRequestBehavior.AllowGet);
@@ -103,7 +103,7 @@ namespace TiendaMascotas.Controllers
   
         [HttpGet]
         //[FiltroSesion]
-        public ActionResult ActualizarProducto(long q)
+        public ActionResult ActualizarProducto(int q)
         {
             try
             {
